change sermon id from int to long

+++ src/main/java/org/improving/dao/SermonRepository.java
@@ -9,11 +9,11 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
 @CrossOrigin( "http://localhost:4200")
-public interface SermonRepository extends JpaRepository< Sermon, Integer >
+public interface SermonRepository extends JpaRepository< Sermon, Long >
 {
     @Transactional
     @Modifying
     @Query("update Sermon sermon set sermon.archiveResource = :archiveResource where sermon.id = :id")
-    void updateArchiveResourceById( @Param("id") int id, @Param("archiveResource") String archiveResource);
+    void updateArchiveResourceById( @Param("id") long id, @Param("archiveResource") String archiveResource);
 
 }
